maven plugin configuration changes artifact renaming to conform to dataflow package renaming to conform to dataflow
Resolves #13
