+++ spring-cloud-starter-dataflow-metrics-collector/src/test/java/org/springframework/cloud/dataflow/metrics/collector/CollectorApplication.java
@@ -1,0 +1,14 @@
+package org.springframework.cloud.dataflow.metrics.collector;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Import;
+
+@SpringBootApplication
+@Import(MetricsCollectorConfiguration.class)
+public class CollectorApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CollectorApplication.class, args);
+	}
+}
