--- spring-cloud-starter-stream-collector-metrics/src/test/java/org/springframework/cloud/stream/app/metrics/collector/CollectorApplication.java
@@ -1,14 +1,0 @@
-package org.springframework.cloud.stream.app.metrics.collector;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Import;
-
-@SpringBootApplication
-@Import(MetricsCollectorConfiguration.class)
-public class CollectorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CollectorApplication.class, args);
-	}
-}
